Create server package and host restapi and a2a there
