+++ agent/remoteagent/utils.go
@@ -19,8 +19,8 @@ import (
 	"slices"
 
 	"github.com/a2aproject/a2a-go/a2a"
-	"google.golang.org/adk/adka2a"
 	"google.golang.org/adk/agent"
+	"google.golang.org/adk/server/adka2a"
 	"google.golang.org/adk/session"
 	"google.golang.org/genai"
 )
