+++ cmd/launcher/adk/adk.go
@@ -18,8 +18,8 @@ package adk
 import (
 	"github.com/a2aproject/a2a-go/a2asrv"
 	"google.golang.org/adk/artifact"
-	"google.golang.org/adk/cmd/restapi/services"
 	"google.golang.org/adk/memory"
+	"google.golang.org/adk/server/restapi/services"
 	"google.golang.org/adk/session"
 )
