+++ examples/workflowagents/sequentialCode/main.go
@@ -24,8 +24,8 @@ import (
 	"google.golang.org/adk/agent/workflowagents/sequentialagent"
 	"google.golang.org/adk/cmd/launcher/adk"
 	"google.golang.org/adk/cmd/launcher/full"
-	"google.golang.org/adk/cmd/restapi/services"
 	"google.golang.org/adk/model/gemini"
+	"google.golang.org/adk/server/restapi/services"
 	"google.golang.org/genai"
 )
