+++ server/adka2a/executor.go
