+++ server/adka2a/metadata.go
