+++ server/adka2a/parts.go
