+++ server/adka2a/processor.go
