+++ server/adka2a/processor_test.go
