+++ server/restapi/handlers/eval.go
