+++ server/restapi/routers/eval.go
@@ -17,28 +17,32 @@ package routers
 import (
 	"net/http"
 
-	"google.golang.org/adk/cmd/restapi/handlers"
+	"google.golang.org/adk/server/restapi/handlers"
 )
 
-// AppsAPIRouter defines the routes for the Apps API.
-type AppsAPIRouter struct {
-	appsController *handlers.AppsAPIController
-}
-
-// NewAppsAPIRouter creates a new AppsAPIRouter.
-func NewAppsAPIRouter(controller *handlers.AppsAPIController) *AppsAPIRouter {
-	return &AppsAPIRouter{appsController: controller}
-
-}
+// EvalAPIRouter defines the routes for the Eval API.
+type EvalAPIRouter struct{}
 
 // Routes returns the routes for the Apps API.
-func (r *AppsAPIRouter) Routes() Routes {
+func (r *EvalAPIRouter) Routes() Routes {
 	return Routes{
 		Route{
-			Name:        "ListApps",
+			Name:        "ListEvalSets",
 			Methods:     []string{http.MethodGet},
-			Pattern:     "/list-apps",
-			HandlerFunc: r.appsController.ListApps,
+			Pattern:     "/apps/{app_name}/eval_sets",
+			HandlerFunc: handlers.Unimplemented,
+		},
+		Route{
+			Name:        "ListEvalSets",
+			Methods:     []string{http.MethodPost, http.MethodOptions},
+			Pattern:     "/apps/{app_name}/eval_sets/{eval_set_name}",
+			HandlerFunc: handlers.Unimplemented,
+		},
+		Route{
+			Name:        "ListEvalResults",
+			Methods:     []string{http.MethodGet},
+			Pattern:     "/apps/{app_name}/eval_results",
+			HandlerFunc: handlers.Unimplemented,
 		},
 	}
 }
