+++ server/restapi/routers/sessions.go
@@ -17,7 +17,7 @@ package routers
 import (
 	"net/http"
 
-	"google.golang.org/adk/cmd/restapi/handlers"
+	"google.golang.org/adk/server/restapi/handlers"
 )
 
 // SessionsAPIRouter defines the routes for the Sessions API.
